Split `MovieMatchers`
