+++ grakn-test/src/test/java/ai/grakn/test/graql/query/MatchQueryTest.java
@@ -65,68 +65,68 @@ import static ai.grakn.graql.Graql.neq;
 import static ai.grakn.graql.Graql.or;
 import static ai.grakn.graql.Graql.regex;
 import static ai.grakn.graql.Graql.var;
-import static ai.grakn.test.matcher.GraknMatchers.aRuleType;
-import static ai.grakn.test.matcher.GraknMatchers.action;
-import static ai.grakn.test.matcher.GraknMatchers.alPacino;
+import static ai.grakn.test.matcher.MovieMatchers.aRuleType;
+import static ai.grakn.test.matcher.MovieMatchers.action;
+import static ai.grakn.test.matcher.MovieMatchers.alPacino;
 import static ai.grakn.test.matcher.GraknMatchers.allVariables;
-import static ai.grakn.test.matcher.GraknMatchers.apocalypseNow;
-import static ai.grakn.test.matcher.GraknMatchers.benjaminLWillard;
-import static ai.grakn.test.matcher.GraknMatchers.betteMidler;
-import static ai.grakn.test.matcher.GraknMatchers.character;
-import static ai.grakn.test.matcher.GraknMatchers.chineseCoffee;
-import static ai.grakn.test.matcher.GraknMatchers.cluster;
-import static ai.grakn.test.matcher.GraknMatchers.comedy;
+import static ai.grakn.test.matcher.MovieMatchers.apocalypseNow;
+import static ai.grakn.test.matcher.MovieMatchers.benjaminLWillard;
+import static ai.grakn.test.matcher.MovieMatchers.betteMidler;
+import static ai.grakn.test.matcher.MovieMatchers.character;
+import static ai.grakn.test.matcher.MovieMatchers.chineseCoffee;
+import static ai.grakn.test.matcher.MovieMatchers.cluster;
+import static ai.grakn.test.matcher.MovieMatchers.comedy;
 import static ai.grakn.test.matcher.GraknMatchers.concept;
 import static ai.grakn.test.matcher.GraknMatchers.constraintRule;
-import static ai.grakn.test.matcher.GraknMatchers.containsAllMovies;
-import static ai.grakn.test.matcher.GraknMatchers.crime;
-import static ai.grakn.test.matcher.GraknMatchers.drama;
+import static ai.grakn.test.matcher.MovieMatchers.containsAllMovies;
+import static ai.grakn.test.matcher.MovieMatchers.crime;
+import static ai.grakn.test.matcher.MovieMatchers.drama;
 import static ai.grakn.test.matcher.GraknMatchers.entity;
-import static ai.grakn.test.matcher.GraknMatchers.family;
-import static ai.grakn.test.matcher.GraknMatchers.fantasy;
-import static ai.grakn.test.matcher.GraknMatchers.gender;
-import static ai.grakn.test.matcher.GraknMatchers.genre;
-import static ai.grakn.test.matcher.GraknMatchers.genreOfProduction;
-import static ai.grakn.test.matcher.GraknMatchers.godfather;
-import static ai.grakn.test.matcher.GraknMatchers.harry;
-import static ai.grakn.test.matcher.GraknMatchers.hasTitle;
+import static ai.grakn.test.matcher.MovieMatchers.family;
+import static ai.grakn.test.matcher.MovieMatchers.fantasy;
+import static ai.grakn.test.matcher.MovieMatchers.gender;
+import static ai.grakn.test.matcher.MovieMatchers.genre;
+import static ai.grakn.test.matcher.MovieMatchers.genreOfProduction;
+import static ai.grakn.test.matcher.MovieMatchers.godfather;
+import static ai.grakn.test.matcher.MovieMatchers.harry;
+import static ai.grakn.test.matcher.MovieMatchers.hasTitle;
 import static ai.grakn.test.matcher.GraknMatchers.hasType;
 import static ai.grakn.test.matcher.GraknMatchers.hasValue;
-import static ai.grakn.test.matcher.GraknMatchers.heat;
-import static ai.grakn.test.matcher.GraknMatchers.hocusPocus;
+import static ai.grakn.test.matcher.MovieMatchers.heat;
+import static ai.grakn.test.matcher.MovieMatchers.hocusPocus;
 import static ai.grakn.test.matcher.GraknMatchers.inferenceRule;
 import static ai.grakn.test.matcher.GraknMatchers.isCasting;
 import static ai.grakn.test.matcher.GraknMatchers.isInstance;
-import static ai.grakn.test.matcher.GraknMatchers.judeLaw;
-import static ai.grakn.test.matcher.GraknMatchers.kermitTheFrog;
-import static ai.grakn.test.matcher.GraknMatchers.language;
-import static ai.grakn.test.matcher.GraknMatchers.marlonBrando;
-import static ai.grakn.test.matcher.GraknMatchers.martinSheen;
-import static ai.grakn.test.matcher.GraknMatchers.mirandaHeart;
-import static ai.grakn.test.matcher.GraknMatchers.missPiggy;
-import static ai.grakn.test.matcher.GraknMatchers.movie;
-import static ai.grakn.test.matcher.GraknMatchers.movies;
-import static ai.grakn.test.matcher.GraknMatchers.musical;
-import static ai.grakn.test.matcher.GraknMatchers.name;
-import static ai.grakn.test.matcher.GraknMatchers.neilMcCauley;
-import static ai.grakn.test.matcher.GraknMatchers.person;
-import static ai.grakn.test.matcher.GraknMatchers.production;
-import static ai.grakn.test.matcher.GraknMatchers.realName;
-import static ai.grakn.test.matcher.GraknMatchers.releaseDate;
+import static ai.grakn.test.matcher.MovieMatchers.judeLaw;
+import static ai.grakn.test.matcher.MovieMatchers.kermitTheFrog;
+import static ai.grakn.test.matcher.MovieMatchers.language;
+import static ai.grakn.test.matcher.MovieMatchers.marlonBrando;
+import static ai.grakn.test.matcher.MovieMatchers.martinSheen;
+import static ai.grakn.test.matcher.MovieMatchers.mirandaHeart;
+import static ai.grakn.test.matcher.MovieMatchers.missPiggy;
+import static ai.grakn.test.matcher.MovieMatchers.movie;
+import static ai.grakn.test.matcher.MovieMatchers.movies;
+import static ai.grakn.test.matcher.MovieMatchers.musical;
+import static ai.grakn.test.matcher.MovieMatchers.name;
+import static ai.grakn.test.matcher.MovieMatchers.neilMcCauley;
+import static ai.grakn.test.matcher.MovieMatchers.person;
+import static ai.grakn.test.matcher.MovieMatchers.production;
+import static ai.grakn.test.matcher.MovieMatchers.realName;
+import static ai.grakn.test.matcher.MovieMatchers.releaseDate;
 import static ai.grakn.test.matcher.GraknMatchers.resource;
 import static ai.grakn.test.matcher.GraknMatchers.results;
-import static ai.grakn.test.matcher.GraknMatchers.robertDeNiro;
+import static ai.grakn.test.matcher.MovieMatchers.robertDeNiro;
 import static ai.grakn.test.matcher.GraknMatchers.rule;
-import static ai.grakn.test.matcher.GraknMatchers.runtime;
-import static ai.grakn.test.matcher.GraknMatchers.sarah;
-import static ai.grakn.test.matcher.GraknMatchers.sarahJessicaParker;
-import static ai.grakn.test.matcher.GraknMatchers.spy;
-import static ai.grakn.test.matcher.GraknMatchers.theMuppets;
-import static ai.grakn.test.matcher.GraknMatchers.title;
-import static ai.grakn.test.matcher.GraknMatchers.tmdbVoteAverage;
-import static ai.grakn.test.matcher.GraknMatchers.tmdbVoteCount;
+import static ai.grakn.test.matcher.MovieMatchers.runtime;
+import static ai.grakn.test.matcher.MovieMatchers.sarah;
+import static ai.grakn.test.matcher.MovieMatchers.sarahJessicaParker;
+import static ai.grakn.test.matcher.MovieMatchers.spy;
+import static ai.grakn.test.matcher.MovieMatchers.theMuppets;
+import static ai.grakn.test.matcher.MovieMatchers.title;
+import static ai.grakn.test.matcher.MovieMatchers.tmdbVoteAverage;
+import static ai.grakn.test.matcher.MovieMatchers.tmdbVoteCount;
 import static ai.grakn.test.matcher.GraknMatchers.variable;
-import static ai.grakn.test.matcher.GraknMatchers.war;
+import static ai.grakn.test.matcher.MovieMatchers.war;
 import static ai.grakn.util.ErrorMessage.MATCH_INVALID;
 import static ai.grakn.util.Schema.MetaSchema.RULE;
 import static java.util.stream.Collectors.toList;
