+++ grakn-test/src/test/java/ai/grakn/test/graql/query/QueryBuilderTest.java
@@ -34,7 +34,7 @@ import org.junit.rules.ExpectedException;
 import static ai.grakn.graql.Graql.insert;
 import static ai.grakn.graql.Graql.match;
 import static ai.grakn.graql.Graql.var;
-import static ai.grakn.test.matcher.GraknMatchers.containsAllMovies;
+import static ai.grakn.test.matcher.MovieMatchers.containsAllMovies;
 import static ai.grakn.test.matcher.GraknMatchers.variable;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
